requirements.txt
openai
python-dotenv
requests
tqdm
icecream
scipy
pandas
surpyval
numpy